Sprite class: Added
